js/xpconnect/tests/mochitest/test_bug655297-1.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=655297
-->
<head>
  <title>Test for Bug 655297</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=655297">Mozilla Bug 655297</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
    <form>0</form> <form>1</form> <form>2</form> <form>3</form> <form>4</form>
    <form>5</form> <form>6</form> <form>7</form> <form>8</form> <form>9</form>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 655297 **/

var map = WeakMap();
function f() {
    var paras = document.getElementsByTagName("form");
    for (var i = 0; i < paras.length; i++)
        map.set(paras[i], "ok");
}
function g() {
    var paras = document.getElementsByTagName("form");
    for (var i = 0; i < paras.length; i++) {
        if (map.get(paras[i]) != "ok") {
            return false;
	}
    }
    return true;
}

f();
SpecialPowers.forceGC();
ok(g(), "Failed to keep XPCWrappedNative used as WeakMap key alive.");

</script>


</pre>
</body>
</html>
